backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.2
psycopg[binary]>=3.2
